fb: override screen_info ioctl with HDR resolution
overriding the FB default implementation of the ioctl-FBIOGET_VSCREENINFO.
This is the standard way to query the driver's display panel resolution.

We wouldn't want to update the device tree according,
because that way, the driver wouldn't be initialized correctly.
and non-expected behaviour will occur.

This way the information inside the driver remains OK,
while notifing a different resolution to the higher
levels (namely - hwcomposer used by the surfaceflinger).

the framework aligns to those initial resolutions,
and draws our explicit resolution.

Change-Id: I6dce6f15e677936b0b36d6e5a2c39a8ea5c7aeab
Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>