mdss_mdp_overlay.c: Add hard-coded pipe offsets depending on mixer id.
Here, we move the drawing rectangle of the left display to the bottom-left corner,
 and the right drawing rectangle to the top-right corner.

Pipes are software components connecting an img-buff to a dma-memory region.
The method mdss_mdp_overlay_pipe_setup() is the one configuring a src_rect and dst_rect.
On each pipe's creation we add custom offset.
In future implementation this values will be dynamically updated.

This way we use the qcom existing driver functions (pipe.offset).
In addition this modification does not require additional processing time.

Change-Id: I213b94da26357d9326c5e38171a48ca58b7b71b6
Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>