mdss_mdp_overlay.c: Forcing all pipes to fetch a full buffer.
Doubling the pipe src_rect and dst_rect inorder to show the full FB on each frame.
forcing display-splitting(user-space) with this commit raise the 2D mirroring feature.

As before while initializing mdss-pipes.
This time we force them to hold double-width data,
and spit it to a double-width target. thus reaching both mixers with
 the same data.

This way we reached a 2D mirror functionality.
this hack is a bit messy, but it is required in order to
avoid modifications in the coupled user-space lib.

affects: this commit does not affect the cursor fb updates.
cursor updates have different data path.

Change-Id: I18db5293f2728b282f94ab9b8c9252cf29154186
Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>