msm8974-mdss.dtsi: disable mdss-fb-split property.
Disabling mdss-fb-split property in display panel deviceTree.
This commit allows the driver to split fb evenly to each display.
Left half to the left-panel, and right fb to the right display-panel.

These values are used when a standard extended display is in used.
for example, in a standard case, when two display panels are connected,
we wouldn't like the FB to split in the middle, rather split only after the
left display-panel is fully used. Thus no black stripes will apear.

In our case, we would like the FB to split exactly in the middle.
Furthurmore we would like black-stripes to appear, and being
able to control them.

this commit impacts a both the driver and the matching user-space lib.
which query this property. see relevant patch in the user-space lib.

Change-Id: I90c4ea5e614ec1df8b5792abba842f6a350378ea
Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>