mdss: add wled calibration sysfs entries
added a global driver pointer to a struct which maps
brightness values of [0,255] to a vector of 3 rgb values.
the index of the array the the key and the structure will
contain the rgb values. sysfs recieves the format:
 * [0,255] [0,1023] [0,1023] [0,1023].
which is parsed as:
 * [brightness-value] [red-level] [green-level] [blue-level].

a userspace script will push values. on-fs mounted.

Change-Id: Id9d041467ad9784e185e97b990da5abcc1b6d5d4
Signed-off-by: Ohad Moisi <ohad.moisi@tandemg.com>